frontend/backend/requirements.txt
fastapi
uvicorn
openai
pillow
python-multipart
pymongo
python-dotenv
pydicom
numpy
sentry-sdk
slowapi
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
open_clip_torch
transformers
